+++ packages/flame/test/assets/images_test.dart
@@ -15,7 +15,43 @@ class MockImage extends Mock implements Image {
 }
 
 void main() {
+  // A simple 1x1 pixel encoded as base64 - just so that we have something to
+  // load into the Images cache.
+  const pixel =
+      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJA'
+      'AAAAXNSR0IArs4c6QAAAA1JREFUGFdjWP33/n8ACGUDhwieHSEAAAAASUVORK5CYII=';
+
   group('Images', () {
+    test('load image', () async {
+      final cache = Images();
+      final image = await cache.fromBase64('img', pixel);
+      expect(image, isA<Image>());
+      expect(cache.fromCache('img'), image);
+    });
+
+    test('access non-existent image', () {
+      final cache = Images();
+      expect(
+        () => cache.fromCache('image'),
+        failsAssert(
+          'Tried to access an image "image" that does not exist in the cache. '
+          'Make sure to load() an image before accessing it',
+        ),
+      );
+    });
+
+    test('access non-yet-loaded image', () {
+      final cache = Images();
+      cache.fromBase64('image', pixel); // did not await
+      expect(
+        () => cache.fromCache('image'),
+        failsAssert(
+          'Tried to access an image "image" before it was loaded. '
+          'Make sure to await the future from load() before using this method',
+        ),
+      );
+    });
+
     test('clear', () {
       final cache = Images();
       final image = MockImage();
@@ -52,11 +88,21 @@ void main() {
 
     test('throws when setting an invalid prefix', () {
       final images = Images();
-
       expect(
         () => images.prefix = 'adasd',
         failsAssert('Prefix must be empty or end with a "/"'),
       );
+    });
+
+    test('.ready()', () async {
+      final images = Images();
+      images.fromBase64('image1', pixel);
+      images.fromBase64('image2', pixel);
+      expect(() => images.fromCache('image1'), failsAssert());
+      expect(() => images.fromCache('image2'), failsAssert());
+      await images.ready();
+      expect(images.fromCache('image1'), isNotNull);
+      expect(images.fromCache('image2'), isNotNull);
     });
   });
 }