+++ packages/flame_tiled/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.3.0-releasecandidate.6
+
+ - Update a dependency to the latest release.
+
+## 1.3.0-releasecandidate.5
+
+ - Update a dependency to the latest release.
+
 ## 1.3.0-releasecandidate.4
 
  - Update a dependency to the latest release.